NO TOUCH ISSUES WERE DISCOVERED. If you have such an issue please contact me by forum or mail.
Fixed an issue with a capital letter inclusion that could cause an issue on certain compilers.
Fixed an issue with cutting engine.
Fixed an issue with fixture properties that were edited in LH not being recognised as integer values.
Fixed an issue with postInit not getting called on dynamically created sprites.